refactor: service와 controller 분리
